server/requirements.txt
Flask
python-dotenv
requests
Flask-Cors
google-auth